move ResourceCloser interface to common
