+++ failgood/common/src/failgood/internal/ResourcesCloser.kt
@@ -1,0 +1,15 @@
+package failgood.internal
+
+import failgood.ResourcesDSL
+import failgood.TestDSL
+import failgood.TestResult
+
+internal interface ResourcesCloser : ResourcesDSL {
+    fun addAfterEach(function: suspend TestDSL.(TestResult) -> Unit)
+
+    fun <T> addClosable(autoCloseable: SuspendAutoCloseable<T>)
+
+    suspend fun closeAutoCloseables()
+
+    suspend fun callAfterEach(testDSL: TestDSL, testResult: TestResult)
+}
